...layout changes

+++ components/fc_config.h
@@ -33,20 +33,21 @@ typedef enum UART_Num{
 	UART_3,
 };
 
+// ...start a all purpose timer by default
 hw_timer_t* timer = NULL;
 
-// ...and there was light
+// ...some calculations
 template<typename T>
 constexpr int MHZ_TO_HZ(T x) {	return ((x) * 1000000); }
 
 template<typename T>
 constexpr int ARRAY_SIZE(T x) {return (sizeof(x) / sizeof(x[0])); }
 
-//
+// ...ESP32 Info
 struct hardware_info {
 	const char* chipModel = ESP.getChipModel();
-	int chipMhz = ESP.getCpuFreqMHz();
-	int chipCores = ESP.getChipCores();
-	int flashChipSpeed = ESP.getFlashChipSpeed();
-	int flashChipSize = ESP.getFlashChipSize();
+	uint32_t chipMhz = ESP.getCpuFreqMHz();
+	uint8_t chipCores = ESP.getChipCores();
+	uint32_t flashChipSpeed = ESP.getFlashChipSpeed();
+	uint32_t flashChipSize = ESP.getFlashChipSize();
 } hardware_info_s;
